apps/nfc_tool/requirements.txt
pyscard
mysql-connector-python
python-dotenv

